Refactor Parquet Statistics
