+++ velox/dwio/parquet/reader/ParquetData.cpp
@@ -17,7 +17,6 @@
 #include "velox/dwio/parquet/reader/ParquetData.h"
 
 #include "velox/dwio/common/BufferedInput.h"
-#include "velox/dwio/parquet/reader/Statistics.h"
 
 namespace facebook::velox::parquet {
