--- velox/dwio/parquet/reader/Statistics.cpp
@@ -1,113 +1,0 @@
-/*
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "velox/dwio/parquet/reader/Statistics.h"
-
-#include "velox/dwio/common/Statistics.h"
-#include "velox/type/Type.h"
-
-namespace facebook::velox::parquet {
-
-std::unique_ptr<dwio::common::ColumnStatistics> buildColumnStatisticsFromThrift(
-    const thrift::Statistics& columnChunkStats,
-    const velox::Type& type,
-    uint64_t numRowsInRowGroup) {
-  std::optional<uint64_t> nullCount = columnChunkStats.__isset.null_count
-      ? std::optional<uint64_t>(columnChunkStats.null_count)
-      : std::nullopt;
-  std::optional<uint64_t> valueCount = nullCount.has_value()
-      ? std::optional<uint64_t>(numRowsInRowGroup - nullCount.value())
-      : std::nullopt;
-  std::optional<bool> hasNull = columnChunkStats.__isset.null_count
-      ? std::optional<bool>(columnChunkStats.null_count > 0)
-      : std::nullopt;
-
-  switch (type.kind()) {
-    case TypeKind::BOOLEAN:
-      return std::make_unique<dwio::common::BooleanColumnStatistics>(
-          valueCount, hasNull, std::nullopt, std::nullopt, std::nullopt);
-    case TypeKind::TINYINT:
-      return std::make_unique<dwio::common::IntegerColumnStatistics>(
-          valueCount,
-          hasNull,
-          std::nullopt,
-          std::nullopt,
-          getMin<int8_t>(columnChunkStats),
-          getMax<int8_t>(columnChunkStats),
-          std::nullopt);
-    case TypeKind::SMALLINT:
-      return std::make_unique<dwio::common::IntegerColumnStatistics>(
-          valueCount,
-          hasNull,
-          std::nullopt,
-          std::nullopt,
-          getMin<int16_t>(columnChunkStats),
-          getMax<int16_t>(columnChunkStats),
-          std::nullopt);
-    case TypeKind::INTEGER:
-      return std::make_unique<dwio::common::IntegerColumnStatistics>(
-          valueCount,
-          hasNull,
-          std::nullopt,
-          std::nullopt,
-          getMin<int32_t>(columnChunkStats),
-          getMax<int32_t>(columnChunkStats),
-          std::nullopt);
-    case TypeKind::BIGINT:
-      return std::make_unique<dwio::common::IntegerColumnStatistics>(
-          valueCount,
-          hasNull,
-          std::nullopt,
-          std::nullopt,
-          getMin<int64_t>(columnChunkStats),
-          getMax<int64_t>(columnChunkStats),
-          std::nullopt);
-    case TypeKind::REAL:
-      return std::make_unique<dwio::common::DoubleColumnStatistics>(
-          valueCount,
-          hasNull,
-          std::nullopt,
-          std::nullopt,
-          getMin<float>(columnChunkStats),
-          getMax<float>(columnChunkStats),
-          std::nullopt);
-    case TypeKind::DOUBLE:
-      return std::make_unique<dwio::common::DoubleColumnStatistics>(
-          valueCount,
-          hasNull,
-          std::nullopt,
-          std::nullopt,
-          getMin<double>(columnChunkStats),
-          getMax<double>(columnChunkStats),
-          std::nullopt);
-    case TypeKind::VARCHAR:
-    case TypeKind::VARBINARY:
-      return std::make_unique<dwio::common::StringColumnStatistics>(
-          valueCount,
-          hasNull,
-          std::nullopt,
-          std::nullopt,
-          getMin<std::string>(columnChunkStats),
-          getMax<std::string>(columnChunkStats),
-          std::nullopt);
-
-    default:
-      return std::make_unique<dwio::common::ColumnStatistics>(
-          valueCount, hasNull, std::nullopt, std::nullopt);
-  }
-}
-
-} // namespace facebook::velox::parquet
